chore: tidy up
