--- Sources/GDSCommon/Patterns/GDSInformation/GDSCentreAlignedViewModel.swift
@@ -1,15 +1,0 @@
-import UIKit
-
-@MainActor
-public protocol GDSInformationViewModelWithTitleAndBody {
-    var title: GDSLocalisedString { get }
-    var body: GDSLocalisedString? { get }
-}
-
-@MainActor
-public protocol GDSInformationViewModelWithImage {
-    var image: String { get }
-    var imageHeightConstraint: CGFloat? { get } // should not have default
-    var imageWeight: UIFont.Weight? { get }
-    var imageColour: UIColor? { get }
-}
